Reworked how @category(...) is handled
Including categories with the `@category` decorator and the `--include` command line option was not working properly before. Changed the approach to add a `categories` member to the classes or test methods decorated with `@category(...)`, and then to use pytest `pytest_runtest_setup` to check for those categories according to the `--include` and `--exclude` command line options.