feat: reworked some topics into examples
The create/load scenario and report topics made more sense as Jupyter notebook examples due to the sequential nature of the snippets they used. The same may be applied to the data types topic. This should improve the state of the build, though more discussion will be necessary to determine exactly how to fit these examples into the user guide.