fix: gracefully degrade migration assistant with Python 3.10
`Traceback.positions` has been added with Python 3.11 making it impossible to extract column ranges. Hence degraded the migration assistant to work without that information at the cost of possibly missing migrations for calls on the same line of code.

Added a warning to the migration assistant documentation to recommend Python >= 3.11.